refactor: mv priceAuthorityRegistry to vats
Somewhat of a breaking change but it's moving from tools which is not part of the package API
