+++ packages/vats/src/vat-priceAuthority.js
@@ -1,5 +1,5 @@
-import { providePriceAuthorityRegistry } from '@agoric/zoe/tools/priceAuthorityRegistry.js';
 import { Far } from '@endo/marshal';
+import { providePriceAuthorityRegistry } from './priceAuthorityRegistry.js';
 
 /**
  * Vat holding the canonical PriceAuthorityRegistry for looking up prices on any
