+++ packages/vats/test/test-priceAuthorityRegistry.js
@@ -1,10 +1,11 @@
 import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
-import { makeScalarBigMapStore } from '@agoric/vat-data';
+
 import { AmountMath, makeIssuerKit } from '@agoric/ertp';
 import { buildManualTimer } from '@agoric/swingset-vat/tools/manual-timer.js';
-import { E } from '@endo/eventual-send';
-import { providePriceAuthorityRegistry } from '../../../tools/priceAuthorityRegistry.js';
-import { makeFakePriceAuthority } from '../../../tools/fakePriceAuthority.js';
+import { makeScalarBigMapStore } from '@agoric/vat-data';
+import { makeFakePriceAuthority } from '@agoric/zoe/tools/fakePriceAuthority.js';
+import { E } from '@endo/far';
+import { providePriceAuthorityRegistry } from '../src/priceAuthorityRegistry.js';
 
 test('price authority confused stores', async t => {
   const baggage = makeScalarBigMapStore('test baggage');
